requirements.txt
langchain_openai
langchain_community
tiktoken
python-dotenv
pypdf
chromadb